Added utility.errorLogger to write log file to c:\errorLogXMS.txt
stringRequest will return an empty string if it crashes.  This is usually because of the server returning
a 403 Forbidden error.  Not sure why it happens sometimes...still logged in and cookies are good.  Will
return correct response after another try.
 - This was source of some crashing in MediaPortal.

Edited ParseChannelData() to return an empty list of changedChannels if the xml file is blank.
 - Can get a blank xml file if server returns a 403 Forbidden and not channel data

Added CheckForIllegalCrossThreadCalls = false to GUI to allow threads trying to update channels

Made any Catch statements use the new errorLogger.